+++ src/amun/cpu/nematus/encoder_decoder.cpp
@@ -20,7 +20,7 @@ using EDState = EncoderDecoderState;
 EncoderDecoder::EncoderDecoder(const God &god,
 							   const std::string& name,
                                const YAML::Node& config,
-                               size_t tab,
+                               unsigned tab,
                                const Nematus::Weights& model,
                                const Search &search)
   : CPUEncoderDecoderBase(god, name, config, tab, search),
@@ -43,7 +43,7 @@ void EncoderDecoder::Decode(EncOutPtr encOut,
 }
 
 
-void EncoderDecoder::BeginSentenceState(EncOutPtr encOut, State& state, size_t batchSize) {
+void EncoderDecoder::BeginSentenceState(EncOutPtr encOut, State& state, unsigned batchSize) {
   EDState& edState = state.get<EDState>();
   decoder_->EmptyState(edState.GetStates(), SourceContext_, batchSize);
   decoder_->EmptyEmbedding(edState.GetEmbeddings(), batchSize);
@@ -85,7 +85,7 @@ mblas::Matrix& EncoderDecoder::GetAttention() {
 }
 
 
-size_t EncoderDecoder::GetVocabSize() const {
+unsigned EncoderDecoder::GetVocabSize() const {
   return decoder_->GetVocabSize();
 }
